refactor(facilityFilterUtil): Extract some utilities out of mapUtil
